fix #119
